contributing a pull-request running shell script
A bash shell script that you can run with the argument of the pull number and it will pull that pull request from the open-oni repo and run it locally. Requires docker, docker-compose, free default ports for open-oni (8983, 3306, 80), a set of batches in a batch folder indicated in the variable BATCHSRCDIR (I use the batches from the Open ONI Sample Data repo).